requirements.txt
apiview==1.3.25
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
python-dateutil==2.8.1
python-git==2018.2.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
xlrd==2.0.1
xlwt==1.3.0
